split out Vision into VisionAutoBase
